backend/requirements.txt
pandas~=1.5.2
numpy~=1.24.0
uvicorn~=0.20.0
fastapi~=0.88.0
starlette~=0.22.0
python-dateutil~=2.8.2
google-api-python-client~=2.70.0
httplib2~=0.21.0
oauth2client~=4.1.3
pydantic~=1.10.2